rerouted file refs and fixed application logic post refactor
