+++ server/aws/clients/s3Client.mjs
@@ -1,4 +1,7 @@
 import { S3Client } from '@aws-sdk/client-s3';
+import dotenv from 'dotenv';
+
+dotenv.config();
 
 const s3Client = new S3Client({ 
   region: process.env.AWS_REGION,
