+++ server/aws/constants/sqlExpressionTemplates.mjs
@@ -1,6 +1,6 @@
 const SQL_EXPRESSION_TEMPLATES = {
   template: 'SELECT * FROM s3object s WHERE',
-  andOperator: 'AND'
-}
+  andOperator: ' AND '
+};
 
 export default SQL_EXPRESSION_TEMPLATES;
